chore: moved home related components to sds/home
