+++ packages/react/src/experimental/Information/exports.tsx
@@ -2,7 +2,6 @@ export * from "./Activity/exports"
 export * from "./Alert"
 export * from "./Avatars/exports"
 export * from "./Badge"
-export * from "./Communities/exports"
 export * from "./Counter"
 export * from "./Headers/exports"
 export * from "./ModuleAvatar"
