+++ packages/react/src/experimental/Navigation/exports.tsx
@@ -1,7 +1,6 @@
 export * from "./ApplicationFrame"
 export { useSidebar } from "./ApplicationFrame/FrameProvider"
 export * from "./Carousel"
-export * from "./DaytimePage"
 export * from "./Dropdown"
 export * from "./F0TableOfContent"
 export * from "./Header"
