+++ packages/react/src/experimental/Widgets/Content/exports.tsx
@@ -1,7 +1,6 @@
 export * from "./CalendarEvent"
 export * from "./CalendarEventList"
 export * from "./CategoryBarSection"
-export * from "./ClockIn/exports"
 export * from "./Highlights/exports"
 export * from "./IndicatorsList"
 export * from "./ListItems/WidgetAvatarsListItem"
