+++ packages/react/src/sds/home/ClockIn/ClockInControls/helpers.ts
