+++ packages/react/src/sds/home/Communities/Celebration/hooks/useConfetti.ts
