+++ packages/react/src/sds/home/Communities/Celebration/index.tsx
@@ -1,11 +1,11 @@
 import { Skeleton } from "@/ui/skeleton"
 import { useEffect, useRef, useState } from "react"
+import { DateAvatar } from "../../../../experimental/Information/Avatars/DateAvatar"
 import { useReducedMotion } from "../../../../lib/a11y"
 import { EmojiImage } from "../../../../lib/emojis"
 import { Link } from "../../../../lib/linkHandler"
 import { withSkeleton } from "../../../../lib/skeleton"
 import { cn, focusRing } from "../../../../lib/utils"
-import { DateAvatar } from "../../Avatars/DateAvatar"
 import { CelebrationAvatar } from "./components/avatar"
 import { useConfetti } from "./hooks/useConfetti"
 import { EMOJI_MAP } from "./types"
