+++ packages/react/src/sds/home/Communities/Celebration/types.ts
