+++ packages/react/src/sds/home/Communities/HighlightBanner/index.stories.tsx
