+++ packages/react/src/sds/home/Communities/HighlightBanner/index.tsx
