+++ packages/react/src/sds/home/HomeLayout/exports.tsx
@@ -1,0 +1,10 @@
+import { Component } from "@/lib/component"
+import { HomeLayout as HomeLayoutComponent } from "."
+
+export const HomeLayout = Component(
+  {
+    name: "HomeLayout",
+    type: "layout",
+  },
+  HomeLayoutComponent
+)
